Has command (#11)
* Spec has command.

* Implement not found when no pid file exists.

* Implement rest of has command.

* Extract get_pid_from_file.

* Refactor output and messaging.

* Fix exception when running `into` on an exited process.

+++ shtuff.py
@@ -71,6 +71,9 @@ def main():
     add_newline_argument(parser_new)
     parser_new.set_defaults(func=shtuff_new)
 
+    parser_has = subparsers.add_parser('has', help='check to see if target is available to receive commands')
+    parser_has.add_argument('name', help='the name of the shell to send the given command to')
+    parser_has.set_defaults(func=shtuff_has)
 
     args = vars(parser.parse_args())
     if not args:
@@ -96,14 +99,17 @@ def shtuff_into(name, cmd, newline):
 
     pid_file = get_pid_file(name)
     if not os.path.exists(pid_file):
-        print(f"Shtuff target {name} was not found.", file=sys.stderr)
+        print_target_not_found(name)
         exit(1)
 
-    with open(pid_file) as f:
-        pid = int(f.read().strip())
+    pid = get_pid_from_file(pid_file)
 
     with open(get_cmd_file(pid), 'w') as f:
         f.write(cmd)
+
+    if shtuff_process_has_terminated(pid):
+        print_target_not_found(name)
+        exit(1)
 
     os.kill(pid, signal.SIGUSR1)
 
@@ -113,8 +119,27 @@ def shtuff_new(cmd, newline):
 
     spawn_and_stuff(os.environ['SHELL'], cmd)
 
+def shtuff_has(name):
+    pid_file = get_pid_file(name)
+
+    if not os.path.exists(pid_file):
+        print_target_not_found(name)
+        exit(1)
+
+    pid = get_pid_from_file(pid_file)
+
+    if shtuff_process_has_terminated(pid):
+        print_target_not_found(name)
+        exit(1)
+
+    print(f"Shtuff process {name} was found with pid of {pid}.")
+
 def get_pid_file(name):
     return data_dir(f"{name}.pid")
+
+def get_pid_from_file(pid_file):
+    with open(pid_file) as f:
+        return int(f.read().strip())
 
 def get_cmd_file(pid):
     return data_dir(f"{pid}.command")
@@ -165,6 +190,12 @@ def spawn_and_stuff(to_spawn, to_stuff):
 
     p.send(to_stuff)
     p.interact()
+
+def print_target_not_found(name):
+    print(f"Shtuff target {name} was not found.", file=sys.stderr)
+
+def shtuff_process_has_terminated(pid):
+    return get_process_command(pid) != 'shtuff'
 
 if __name__ == "__main__":
     main()
